new file for cross BLST tests
